requirements.txt
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.5.0
azure-identity==1.15.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
azure-storage-file-share==12.21.0
chromadb==1.0.15
gitdb==4.0.12
GitPython==3.1.44
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
openpyxl==3.1.5
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
psutil==7.0.0
PyJWT==2.10.1
pyodc==1.5.0
pypandoc==1.15
pypdf==5.7.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
streamlit==1.46.1
unstructured==0.18.2
unstructured-client==0.37.4
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15